refactor: break UnifiedItemEditModal into smaller components
- Extract QuantityShortcuts component for quantity shortcut buttons
- Extract QuantityControls component for input field and increment/decrement buttons
- Extract UnsupportedItemMessage component for unsupported item type message
- Extract UnifiedItemEditBody component as main body that orchestrates all sub-components
- Update UnifiedItemEditModal to use new component structure
- Improve component modularity and prepare for future support of reference === 'group'
- All tests and type checks pass
