+++ src/sections/unified-item/components/UnsupportedItemMessage.tsx
@@ -1,0 +1,8 @@
+export function UnsupportedItemMessage() {
+  return (
+    <div class="text-gray-400 text-sm">
+      Este tipo de item não é suportado ainda. Apenas itens de comida e receitas
+      podem ser editados.
+    </div>
+  )
+}
